+++ src/UI/Fields/AmountSelect/components/AmountInput/helpers/hasLeadingZeroes.ts
@@ -1,0 +1,14 @@
+import { ZERO } from '@multiversx/sdk-dapp/constants';
+
+export const hasLeadingZeroes = (value: string) => {
+  const [firstCharacter, secondCharacter] = value.split('');
+
+  const firstChacterIsZero = firstCharacter === ZERO;
+  const secondCharacterIsPeriod = secondCharacter === '.';
+
+  if (firstChacterIsZero && secondCharacter && !secondCharacterIsPeriod) {
+    return true;
+  }
+
+  return false;
+};